+++ tests/testthat/test-cache.r
@@ -4,6 +4,10 @@ test_that("works on mac and linux", {
   expect_equal(user_cache_dir("R", os = "mac"), "~/Library/Caches/R")
 })
 
+test_that("can override with R_USER_CACHE_DIR", {
+  withr::local_envvar(R_USER_CACHE_DIR = "/test")
+  expect_equal(user_cache_dir("R", os = "mac"), "/test/R")
+})
 test_that("works on windows simulation", {
   skip_on_os("windows")
   expect_equal(